sessionmgr/client: Save correct backend pid
There was a slight confusion with which PID was stored in the session
manager and which PID the real backend process had.  This was caused due
to the backendstart service reporting the PID it got when starting the
client process, while the client process would do yet another round of
fork() before settling down.

There is value having the initial PID for debuging, if the client
process doesn't manage to fork() or fails before this point.  This
information will the then be in the logs.  But once the client process
has forked and started the registration process with the session
manager, the initial PID have no longer any value.

So the solution is basically to report the running client PID to the
session manager via the registration request signal and update the
backend_pid variable in the session object based on this value.

Signed-off-by: David Sommerseth <davids@openvpn.net>