policy: Grant access to D-Bus Properties.GetAll
Commit f7bc93f0f227d766a5b25 added the CheckObjectExists() method which
uses org.freedesktop.DBus.Properties.GetAll() to get a confirmation if
the object exists or not.  This change caused the session manager to
spew out quite a few access denied errors in the log when starting and
managing backend VPN client processes.

This change is not optimal, as it gives the openvpn user account access
to org.freedesktop.DBus.Properties.GetAll on all services.  But it
silences these warnings for now.

Signed-off-by: David Sommerseth <davids@openvpn.net>