sessionmgr: Save the config name in the session object
For single-use configuration profiles or long lasting sessions where the
configuration profile might be deleted in between, there is no
possibility to retrieve any configuration name hint of which the session
is based on.

This changes the RegistrationConfirmation() call the session manager
does to the backend VPN client process (openvpn3-service-client).  When
the VPN client process retrieves the configuration profile it will now
also retrieve the configuration name and return this string in the
RegirstrationConfirmation() call.  The session manager then saves this
value, which will always be available during the lifetime of the
session.

Signed-off-by: David Sommerseth <davids@openvpn.net>