logger: Add idle-exit logic to the log service
When openvpn3-service-logger is started with --service, it will now run
with an idle-exit timer set to 10 minutes by default.  If no
openvpn3-linux based services have attached to the service in a 10
minutes time window, the logger will exit automatically.

As with all the openvpn3-linux services, this can be modified via the
--idle-exit argument and setting it to 0 will disable the idle exit
timer.

Signed-off-by: David Sommerseth <davids@openvpn.net>