logger: Add --state-dir option to openvpn3-service-logger
Added --state-dir to be used together with --service.  This option takes
a directory name as argument.  When any of the tweakable settings in the
log service is modified, a JSON file is created in this directory
preserving the settings.  When the service is started later on, it
will parse this file and restore the previous settings.

Signed-off-by: David Sommerseth <davids@openvpn.net>