configmgr: Allow root to transfer config profile ownership
This allows the root user on the system to manipulate the owner property
in configuration profile D-Bus objects via a new D-Bus method,
TransferOwnership in the main configuration manager object.

This is hard-coded to be restricted to the root user for security
reasons.  This is a very powerful feature which might be otherwise
abused more easily.  This restriction is implemented both in the C++
code as well as the D-Bus policy - which is fairly simple to control
when having it as a separate method.

This method was also added to the main configuration manager object to
attempt to make it even clearer this is a special feature and not a
common operation.  The main user of this feature will be the
openvpn3-autoload utility, which can be run as part of the boot process
on a system.

Signed-off-by: David Sommerseth <davids@openvpn.net>