sessionmgr: Allow root to transfer session ownership
This is similar to commit d128f1df7a6edee which enables a similar
feature on configuration profile objects.  This patch adds the same
feature on session objects, which allows the root user to change the
session owner D-Bus object property via the TransferOwnership method in
the main session manager object.

This is hard-coded to be restricted to the root user for security
reasons.  This is a very powerful feature which might be otherwise
abused more easily.  This restriction is implemented both in the C++
code as well as the D-Bus policy - which is fairly simple to control
when having it as a separate method.

As with the configmgr, this method was added to the main session manager
object to emphasize this is a special use feature and not intended for
normal management operations, primarily to be used via the
openvpn3-autoload utility which can start VPN sessions automatically
during the system boot.

Signed-off-by: David Sommerseth <davids@openvpn.net>