netcfg: Add global DNS settings in main service object
This is the firrst part of splitting up some information related to DNS
server and search domain details.  The main netcfg service object in
/net/openvpn/v3/netcfg is extended with global_dns_servers and
global_dns_search properties, which should contain the same information
as is active on the system.

With the --resolv-conf/ResolvConfFile class, this typically means the
contents /etc/resolv.conf should be reflected in this main service object.
If one or more virtual interfaces are added, they will also be seen in
these lists.

Signed-off-by: David Sommerseth <davids@openvpn.net>