netcfg: Only report device specific DNS settings on a single interface
When retrieving the dns_servers and dns_search properties of a single
interface, the complete system list was returned.  Since there is now
global dns_servers and dns_search lists which includes the pre-defined
system entries plus the device specific entries, the DNS information of
a single interface should be restricted to that interface only.

This changes the device specific DNS information to be returned in the
dns_servers and dns_search properties of the interface's unique object.

Signed-off-by: David Sommerseth <davids@openvpn.net>