netcfg: Ensure the DNS server list from VPN server is preserved
This caches up DNS server list additions in a separate list which is
added to the main list when calling the new CommitChanges() method.  It
is expected that the implementor of  the ResolverSettings calls this
method at an appropirate place in the Apply() method.  Without this
call, the DNS server list will be unchanged.

This ensures that AddDNSServer() can be called multiple times and the
server list order is preserved, no matter if the AddDNSServer() parses a
list via a GVariant object or a single DNS server per AddDNSServer()
call.

Signed-off-by: David Sommerseth <davids@openvpn.net>