dbus/proxy: Implement support for passing file descriptors
This extends the DBusProxy API with new methods, CallGetFD() and
CallSendFD(), which are used to retrieve file descriptors passed by
the service or send file descriptors to a service.

Please do notice that the file descriptor passing possibility is
not clearly defined in the the D-Bus introspection.

This is based on several fragments written by Arne Schwabe, which
have been extracted and incorporated into this single commit.

Signed-off-by: Arne Schwabe <arne@openvpn.net>
Signed-off-by: David Sommerseth <davids@openvpn.net>