netcfg: Add various protect options on the command line
This implements a new NetCfgOption struct which carries some settings
which can be set via on the command line.  These settings are then
passed on all the way until the NetCfgDevice object.

These options are currently not being used in the NetCfg or
NetCfgDevice objects yet, but this prepares the ground for it.

Signed-off-by: Arne Schwabe <arne@openvpn.net>
Signed-off-by: David Sommerseth <davids@openvpn.net>