configmgr: Move the explicit --persist-tun to an override
This modifies the configuration manager to process --persist-tun as a
configuration override instead of a separately handled object property.

This is a setting which might be site specific, thus having it as
an override makes it easier for the users compared to modifying and
re-importing a configuration profile from file.

Signed-off-by: David Sommerseth <davids@openvpn.net>