dbus/proxy: Extend DBusProxy with CheckServiceAvail() method
This method will check if a service is already running or attempt to
explicitly start it.  If the service was started by this call, it will
currently try to do a Ping() call as well, to ensure the service is
responding.

This can be used as an alternative to calling just Ping().

Signed-off-by: David Sommerseth <davids@openvpn.net>