log: Extend LogConsumer to process other signals as well
It might be wanted to handle more signals than just the Log signal in
LogConsumer, but the implementation of the signal handling in this case
needs to be implemented externally.

This extends the LogConsumer with a new ProcessSignal() method which can
be overridden.  All signals this object has subscribed to which are not
Log signals will be sent to this callback method instead.

Signed-off-by: David Sommerseth <davids@openvpn.net>