cli/log: Check if session exists before restoring settings
When the 'openvpn3 log' command was about to stop running, it would
attempt to reset the receive_log events flag if it changed it.  But the
session itself might have been closed, so there is no available session
object.

Add an additional check to see whether the D-Bus object still exists
before attempting to restore this setting.

Signed-off-by: David Sommerseth <davids@openvpn.net>