sessionmgr/proxy: Implement LookupConfigName() for sessions
This is a similar C++ interface implemented in the configuration manager,
but for the session manager.  This lookup method will retrieve a list of
running VPN sessions which where started with a specific configuration
profile name.  This lookup happens against the net.openvpn.v3.sessions
D-Bus service.

The unit test which targeted the configuration manager has been extended
and renamed, as it can be reused for both the configuration and session
manager.

Signed-off-by: David Sommerseth <davids@openvpn.net>