sessionmgr: Implement LookupConfigName() method
This implements a lookup method of configuration names and returns a
list of running VPN sessions the caller has access to.  This is
essentially the same lookup approach used in the configuration manager,
but targetting the session manager and running VPN sessions instead.

Since the current implementation does not change the configuration
profile name in running VPN sessions if the configuration changes after
the session has started, this lookup method will use the configuration
profile name in use when the session was started.

Signed-off-by: David Sommerseth <davids@openvpn.net>