client: Use more unique device name than 'o3tun'
When creating a virtual interface via the netcfg service, it used a
hard-coded 'o3tun' device name which made little sense.  This switches
to use the session token, provided on the command line.  The D-Bus
device path will from this have both the VPN client's PID + this
session token which is quite close to guaranteed to be unique.  And
it makes it easier to provide information to more easily map the
VPN client process to a specific device.

Signed-off-by: David Sommerseth <davids@openvpn.net>