openvpn3: Revamp the code preparing the available commands
Change the approach of declaring new commands by not depending
on access to the Commands object.  Rather flip this around so
there are "preparation" functions which returns a SingleCommand
object which is registered into the Commands object instead.

This allows a greater flexibility and we can later on move
towards an even more proper object oriented approach for the
SingleCommand implementations.

Signed-off-by: David Sommerseth <davids@openvpn.net>