log: Extend LogEvent to be more self-contained
This puts all the D-Bus related processing of Log signals into the
LogEvent class, so all the code related to create and parse GVariant
objects with LogEvents can be shared, as well as comparison operators.

This also extends the GVariant object parser to process both tuple and
dictionary based input data.

A new unit-test program for the LogEvent class is also added.

Signed-off-by: David Sommerseth <davids@openvpn.net>