log: Extend LogEvent to support session-tokens
The back-end client process will need to provide some additional unique
identifier, a session token.  This can't be extracted from the D-Bus
object path as that needs to be static to allow a more stricter D-Bus
policy.

This extends the LogEvent class to parse and generate D-Bus data which
contains this additional information.

Signed-off-by: David Sommerseth <davids@openvpn.net>