sessionmgr/log: Allow sessionmgr to intercept and manipulate proxied LogEvents
This changes the LogConsumerProxy class to give the implementor a chance
to manipulate the LogEvents before being proxied.  This is used to
remove the session_token value from front-ends, which is not needed here
and which would otherwise break the signal signature when this value is
provided in the LogEvent from the VPN client backend.

Signed-off-by: David Sommerseth <davids@openvpn.net>