exceptions: Clean up, simplify and unify
The exception classes was quite diversive and at times consuming in the
API.  In addition, several of them had multiple methods to retrieve the
same information, often only variying between std::string and
const char * as the return type.  Some exception classes was also
needlessly complicated.

This cleans up all this and tries to strive for a more uniform and
simpler API.  It also prefers to use const char * as the return type
used to retrieve the error string, as well as using the C++ standard of
using the what() method wherever possible.  A few places a more
specialized approach is needed, or it is needed to have additional
methods to get other kind of details.

Signed-off-by: David Sommerseth <davids@openvpn.net>