log: Extend LogConsumerProxy::InterceptLogEvent() with error handling
It was no possibility for the InterceptLogEvent() implementation to
indicate failures in a safe way without causing a bigger error.  This is
needed to tell if the ProxyLog() should be skipped or not.  There are
situations where the InterceptLogEvent() implementation may want to
filter out log events.

This also enables a more flexible scheme which can easily enable
extending this further to handle other scenarios as well.

Signed-off-by: David Sommerseth <davids@openvpn.net>