sessionmgr: Add filtering on StatusChange when proxying
In commit b2830e7aee2aad, the backend VPN client processes switched over
to a static object path, as the session token, pid and unique bus name
will still be different.  This gave a side-effect in the
SessionStatusChange class, where it started to proxy signals from
VPN sessions unrelated to this specific object.

To avoid extending the StatusChange signals to also carry a session
token value, it was considered better to just do a match against the
unique bus name of the sender.  This ensures the ProxyCall() method will
only be called whenever the the StatusChange signal comes from the same
backend VPN client process this SessionStatusChange object is configured
for.

Signed-off-by: David Sommerseth <davids@openvpn.net>