client: Enable compile time session manager check disabling
When the openvpn3-service-client process starts, it only allows the
session manager to call methods in the backend VPN client D-Bus object.
This is to ensure that all backend processes are tightly coupled with
the session manager and it reduces the attack vector.

This restriction makes some deeper going debugging impossible, where the
developer may need to call these D-Bus methods directly.

This is solved by adding a --enable-debug-disable-sessionmgr-check
argument to ./configure, which basically allows any callers to call
these methods - as long as the D-Bus policy also allows it.

This is considered a security weakening so it should not be enabled on
production builds.  This is targeting advanced debug sessions.

Signed-off-by: David Sommerseth <davids@openvpn.net>