client: Fix crash when dynamic challenge auth is required
The openvpn3-service-client process could crash (SIGABRT) if the server
sends a DYNAMIC_CHALLENGE back to the client.  This requires the client
to shutdown, acquire the needed input for the authentication and start a
new connection.

What happened was that the client_thread was not properly cleaned up
before starting a new connection, which caused memory pointers to the
thread object to go astray.

This is fixed by ensuring we complete the client_thread by joining it
and nullify it before creating a new thread.

Signed-off-by: David Sommerseth <davids@openvpn.net>