configmgr: Remove support for aliases
It was a partially functional functionality allowing configuration
profiles to be accessible via aliases.  This implementation was buggy
and unstable and after the support for using --config instead of the
D-Bus configuration path (via --path/--config-path), the alias feature
was even less usefull.

So rather rip this code out, as it is not really interesting anymore.
If it will become useful later on, it will be better to implement it
from scratch.  Having a tidy code base is more important now.

Signed-off-by: David Sommerseth <davids@openvpn.net>