configmgr/json: Clean up OptionListJSON class
The json_export() method returned a std::string.  Switch this to
Json::Value instead and add the string conversion where it truly is
needed.  This allows for a simpler implementation later on.

Also remove the FIXME related to the for-loop iteration while tightening
it up a bit by using const auto&.  Applies the same change to the
string_export() method as well.  This seesm to work quite well and seems
to be a reasonable approach.

Signed-off-by: David Sommerseth <davids@openvpn.net>