dbus/proxy: Add retry logic in DBusProxy::CheckObjectExists()
With the introduction of the persistent storage in the configuration
manager, objects may exists but might not be available at that moment
due to the service providing it has not managed to start and settle yet.

By adding a tunable to how many times it should try before considering
it a fail as well as modifying how long it should wait before each
attempt, this misbehaviour is possible to avoid.

This implementation uses a default value of 1 attempt only and with a
default of 1ms delay between each attempt.  Existing implementations
does not need to change anything, and the behaviour will in this
situation be identical with the prior behaviour.

There is also a slight refactoring happening, to avoid to much needless
if() nesting.

Signed-off-by: David Sommerseth <davids@openvpn.net>