ovpn3cli: Be graceful to configmgr if it needs to start
After the configuration manager added support for persistent storage,
old configuration paths of persistent configuration files will be the
same.  But when 'openvpn3 sessions-list' wants to query the config
manager for details, the configuration manager may not be running.

Without this patch, sessions-list would just report "(Config not
available)" if the configuration manager had to start.  But on the next
attempt shortly after, it would not report the configuration as
unavailable.

This change makes use of the the extended capabilities in
DBusProxy::CheckObjectExists() to try more times before considering the
configuration being unavailable.  This is set to 100 attempts for now.
This gives the configuration manager at least 100ms to load all the
persistent configuration profiles and complete initializing.  Normally
4-5ms is enough, including loading 3 persistent configuration profiles.

Signed-off-by: David Sommerseth <davids@openvpn.net>