configmgr/docs: Misc updates to the service documentation
The D-Bus service documentation for the configuration manager was
missing a few details as well as there were some syntax issues with the
markup.

Added information about the version property as well as the
configuration overrides, plus re-ordered the properties list to be
better aligned with the currently used order.

Signed-off-by: David Sommerseth <davids@openvpn.net>