common: Prepare SingleCommandOption to handle alias arguments
This extends get_struct_option() and gen_help_line() to return
std::vector of their previous return types and does the needed
adjustments to the SingleCommand class to handle this change.

This will later on allow a single SingleCommandOption to handle more
arguments as a single argument, typically useful for alias arguments.

Signed-off-by: David Sommerseth <davids@openvpn.net>