dbus: Extended GLibUtils with ExtractValue()
This can be used as an alternative to g_variant_get(), together with
checkParams().  This uses a different approach than g_variant_get().

The g_variant_get() function in glib2 can be troublesome in some
cases; it can easily leak memory and occasionally crash when
freeing memory allocated by this function.  Why this happens some times
is not much clear, and it seems to be highly dependent on the glib2
version and compiler being used, as well as compiler optimization
flags.  We need something more robust and to avoid random crashes.

Signed-off-by: David Sommerseth <davids@openvpn.net>