client: Auto-detect if client certificate is present
Until this change, it was needed to manually add the --no-client-cert
override to the configuration file if a configuration profile did not
provide any client certificate.  This should be detected automatically
and set the needed flags; this makes the client behave more like the
OpenVPN 2.x generation.

This adds the needed detection (which is fairly comprehensive) in the
openvpn3-service-client code and removes the --no-client-cert override
all the other places and ensures the documentation is up-to-date.

A big thanks goes to S. Saurymper who reported this issue and helped
with lots of debugging and testing for this patch.

Signed-off-by: David Sommerseth <davids@openvpn.net>