netcfg: Fix unexpected DNS resolver setup restore
Quite often the DNS resolver settings where restored back to the
starting point long before it should when running multiple VPN sessions
and disconnecting one or two of them.

The reason was that the reference counter was not properly initialized
which would then result in the counter reaching 0 on the wrong condition
instead of after the last VPN session closing.  This resulted in the
Restore() method being executed instead of the Apply() method.

Fix is simple enough: Ensure the reference counter being initialized
correctly.

Signed-off-by: David Sommerseth <davids@openvpn.net>