netcfg: Don't try to restore DNS without active backup
If connecting to a VPN server and the DNS resolver settings has not been
activated, there will not be any backup settings to restore; it will
only be created if the settings has changed.

Without this change, netcfg will complain about backup file missing.
In this case, this is not an error but expected.  Instead exit early if
we know no backup file has been created.

Signed-off-by: David Sommerseth <davids@openvpn.net>