ovpn3cli/sessions: Add --cleanup to session-manage
It may happen that openvpn3-service-client processes dies or does not
report any sensible status back to the session manager.  This results in
the session manager keeping a list of non-functional sessions.

This cleanup feature will walk through all registered sessions and
validate if they are still properly running.  Those not being valid will
be cleaned up and removed.

Signed-off-by: David Sommerseth <davids@openvpn.net>