sessionmgr: Improve backend Ping() error handling when registering
On some slower responding systems it could appear situations where the
session manager's Ping() call to the backend VPN client process could
fail as it would not have settled complete after sending the
RegistrationRequest signal.

The solution here is simply to be more gracious if errors occurs when
trying to Ping() the backend process.  Try up to 5 times, with a sleep
of 1 second between each iteration.

Signed-off-by: David Sommerseth <davids@openvpn.net>