dbus: Revamp IdleCheck to use std::condition_variable
The prior implementation was rather simplistic and was based around a
simple sleep timer.  This had several issues, especailly when sending
SIGINT or SIGTERM signals, as it could often take some time before it
really exited.

This completes the FIXME comment suggesting switching over to
conditional variables which has a much better interface for our use
case.  This enables exiting the "sleep" instantly when the condition
changes, which results in an instant termination of the program.

One important change here is that IdleCheck will now setup the signal
handling itself.

Signed-off-by: David Sommerseth <davids@openvpn.net>