build: Strip out build date/time stamp by default
This will remove the use __DATE__ and __TIME__ unless the OPENVPN_DEBUG
macro is enabled; which gets enabled via ./configure
--enable-debug-options.

This also renames DEBUG_OPTIONS to OPENVPN_DEBUG.  This align the
behaviour to be similar to the OpenVPN 3 Core Library.  The Core Library
is also updated to pull in the reproducible build commit stripping out
the same kind of timestamp unless OPENVPN_DEBUG is defined.

Signed-off-by: David Sommerseth <davids@openvpn.net>