client/netcfg: Add proper support for persist-tun
Until this change, the persist-tun feature has not been too functional.
This has been a tricky fix, as it may need to reconfigure the bypass
gateway rule when this is used together with --redirect-gateway.

The needed support for this has finally landed properly in the Core
Library, so this change also pulls in the latest Core Library update as
well.  And in the Linux client, we just need to enable and configure
the needed hooks to resolve this issue.

The downside with this fix is that it may reconfigure a new TUN device
before swapping the trafic over to the new OpenVPN server.  This is to
ensure no data is unexpectedly leaked outside the tunnel.

Signed-off-by: David Sommerseth <davids@openvpn.net>