tests: Adding unit-test framework
This adds a unit-test framework based on googletest (gtest).  Right now
only the bare framework is introduced, not running any particular tests.

This also moves the D-Bus policy syntax check to the main Makefile.am to
avoid running multiple test scripts via 'make', which did look a bit
confusing.  This required a slight modification to the syntax-check.sh
script, where a lacking copyright header was added as well.

In addition, the BUILD_TESTS declaration has been changed to
BUILD_TEST_PROGS, as it is possible to build unit-tests without the
additional test programs and vice versa.  The default is to build
everything.

Signed-off-by: David Sommerseth <davids@openvpn.net>