client: Tackle better if NetCfgProxy object is lost
D-Bus can be pretty crude if something bad happens, where it just
disconnects the proxy from the bus and service it is tied to.  This can
happen if an unexpected error occurs inside D-Bus.

This can happen if the CoreClient's TunBuilder tries some D-Bus calls
which fails, which further results in the device object not being
accessible or valid any more.  These additional checks tries to avoid an
random crash and instead be able to exit more gracefully.

Signed-off-by: David Sommerseth <davids@openvpn.net>