log: Don't throw access denied exception during init
The constructor calls the DBusProxy::GetServiceVersion() method during
the object initialization.  This makes everything misbehave if the user
running this code does not have access to the 'version' property in the
net.openvpn.v3.log service.  This makes no sense, as this check is here
primarily to ensure the service is available and running.

By just ignoring any DBusProxyAccessDeniedException events, we can
continue without explosions and further access denials will be handled
later on any how.  The check here is primarily to ensure the service is
available.

Signed-off-by: David Sommerseth <davids@openvpn.net>