tests: Fix failing unittest with --enable-debug-core-events
When ./configure --enable-debug-core-events was used, the
StatusEvent.stringstream test failed as it included the numeric status
codes in the output.  This is because DEBUG_CORE_EVENTS sets this to be
the default in StatusEvent objects.  So we just ensure this feature is
disabled if DEBUG_CORE_EVENTS is in use for the failing test; which
restores the expected behaviour.

Signed-off-by: David Sommerseth <davids@openvpn.net>