sessionmgr: Expose NetCfg device path and name for VPN sessions
This extends the session objects with both a D-Bus path to the netcfg
device configuration (handled by the openvpn3-service-netcfg service) as
well as the device name currently used by this VPN session.  This
information is pulled directly from the backend VPN client process
(openvpn3-service-client).

This also extends the proxy interface for the session manager to have an
easy access to the device name.  Currently it is not believed the netcfg
device path is needed in the proxy interface, as this is more a
lower-level debug detail.

Signed-off-by: David Sommerseth <davids@openvpn.net>