core-ext: Replace hacky option string export
The OptionListJSON::string_export() method had a fairly hacky
implementation to handle the static-challenge option.  The Option object
returned in the for-loop iterator has the escape() and render() method.

Unfortunately, they don't handle inline files properly.  So we use the
escape() method for non-inline files and the local optparser_mkline()
function to handle the inlined file data.

This is still not optimal, but seems to be better than the earlier
implementation.

Signed-off-by: David Sommerseth <davids@openvpn.net>