dbus/proxy: Strip "remote" part of GDBus errors
The glib2/gdbus documentation recommends to call
g_dbus_error_strip_remote_error() on errors before presenting errors to
the user.  This is typically needed in the D-Bus proxy code, as this is
where a program is calling a "remote" service.

Since the error is somewhat parsed a few places, we now need slightly
modify the string matching to remove the remote references as well.

Signed-off-by: David Sommerseth <davids@openvpn.net>