sessionmgr/proxy: Verify session objects availability before calling it
If the front-end caller is too fast, it may hit a race condition where
the session object has not managed to establish itself before the
front-end tries to access it.  This adds an additional check to ensure
both the session manager as well as the session objects are available
before performing method calls against them.

Signed-off-by: David Sommerseth <davids@openvpn.net>