dbus/idlecheck: Consider last_operation when executing idle exit
Without this fix, the code only considered the reference counter to
consider if an application was idle or not.  This is enough in many
cases, but if there happens to be equal amount of reference
increment and decrements within the idle check time window, the
reference counter would be 0; thus the program was considered to
be idle while it had been in use during this time window.

This adds a check to check whether the last_operation timestamp
has exeeded the time window for the idle check.  And only if this
is true it will consider the reference counting and act accordingly.

Signed-off-by: David Sommerseth <davids@openvpn.net>