tests: Add simple connect/disconnect stress test
Include a simple test script to connect and disconnect to a remote VPN
server in a loop.  This is useful to test various race conditions and
the basic stability of the OpenVPN 3 Linux client.

This also includes a simple configuration template for both the client
profile and the server configuration.

Signed-off-by: David Sommerseth <davids@openvpn.net>