dbus/signals: Add getter methods for interface and object path
Add two new getter methods to retrieve the D-Bus object interface and
object path a signal subscription is attached to.

Since the method names are fairly universal, this will cause compiler
confusion in cases where both DBusObject and DBusSignalSubscription
classes are inherited in parallel.  This results in needing to specify
the proper class base to use in SessionObject.

Signed-off-by: David Sommerseth <davids@openvpn.net>