ovpn3cli/session: Minor overhaul of sessions-list
The sessions-list command could be a bit fragile at times if querying
the session manager right in between a new tunnel was started but before
it had completed the initialization.  This caused needless error
messages instead of gracefully ignoring data it could not retrieve at
the time of the run.

In addition the overall formatting has received a needed overhaul, to
ensure proper alignment and that contents of unavailable data is
somewhat consistent.

Signed-off-by: David Sommerseth <davids@openvpn.net>