client: Provide more information on non-sessionmgr rejections
If a D-Bus caller is not identified as the session manager, it would
reject the call with very little details about the caller.  Since
this is logged as well, the log information would not help identify
the caller.

This extends the rejection to include the unique bus name of the
caller, together with PID and UID.

Signed-off-by: David Sommerseth <davids@openvpn.net>