netcfg: Add a new DNS settings framework
The current DNS configuration framework is too limited to work well with
systemd-resolved and has some duplication issues with the current ResolvConf
implementation.

This adds a new framework which has a better split between the handling of
DNS settings and the backend part which will perform the tasks on the
system.  The intention is that each device will have their own
ResolverSettings object which the DNSSettingsManager will pass to the
ResolverBackendInterface implementation.

When a device is removed, the ResolverSettings object is also removed and
on the DNSSettingsManager::ApplySettings() is called, only the active
settings are processed.  This also enables extending the
ResolverSettings to also pass needed device references which some
ResolveBackendInterface implementations may need.

Signed-off-by: David Sommerseth <davids@openvpn.net>