netcfg: Implement new resolv.conf backend processor
This will replace the current src/netcfg/dns-direct-file.hpp
implementation for processing and managing the resolv.conf file on the
system.

This change only implements a reworked FileGenerator and ResolvConfFile
classes, based on the new ResolverBackendInterface.  When this new
implementation is taken into use, the old implementation will be
removed.

The self-test program is adopted as well, but should be rewritten into a
more suitable unit-test format.

Signed-off-by: David Sommerseth <davids@openvpn.net>