build: Updating default C++ compiler flags
The default compiler flags removed quite a few warnings, which were not
strictly required.  But we do add -Wno-maybe-uninitialized due do a
warning from the Core library where this issue is reported.  This is
believed to be a false-positive for the time being.  Once the Core
library has resolved this issue, then we can re-enable this warning
again.

Also remove -Wl,--no-as-needed which is believed is not strictly needed.
This should be the default behaviour anyhow.

This also fixes incorrect usage of AM_LDFLAGS.  LDADD is what should be
used, and this was discovered when removing the --no-as-needed linker
flag.

Signed-off-by: David Sommerseth <davids@openvpn.net>