netcfg/signals: Add support for providing device name with Debug logging
This adds an overload method which takes the device name as an
additional argument to the Debug() method.

Note that the device_name value in the NetCfgDevice object may not be
set by the Core library yet when logging, so it might somewhat cryptic
until Core library knows the real device name.  In the mean time it
carries the name which is part of the D-Bus object path.

Signed-off-by: David Sommerseth <davids@openvpn.net>