netcfg: Add a mutex lock for changes to the resolv.conf file
On a heavy loaded system, it may take some time before the change is
done - which in the mean time could mean another VPN session tries to
modify the resolv.conf settings.  So add a lock guard early in the
Commit() method to ensure the file is not modified by another VPN
session at the same time.

Also wipe the unprocessed_lines when we're done writing the resolv.conf
contents to disk; to avoid another potential duplication of these
unprocessed system settings.

Signed-off-by: David Sommerseth <davids@openvpn.net>