ovpn3cli/netcfg: Add presenting PID of subcribers in --list-subscribers
When listing netcfg subscribers, it now only presents the unique D-Bus
name for each subscription.  It requires a bit of additional D-Bus
lookups to figure out which process the subscription belongs to.

The DBusConnectionCreds class can easily extract the information we
need, implement the GetPID() lookup and provide the PID of the
process in the subscriber listing.

Signed-off-by: David Sommerseth <davids@openvpn.net>